Fix difficulty adjuster type conversions from u64 to f64
- Update DifficultyAdjuster to handle f64/u64 conversions properly
- Fix account_for_unaccounted_shares to return u64
- Fix apply_difficulty_constraints to accept u64 and return f64
- Update set/get current_difficulty methods to use f64
- Ensure type consistency across difficulty calculations

+++ p2poolv2_lib/src/accounting/calc.rs
@@ -44,7 +44,7 @@ pub fn time_bias(time_difference: f64, period: f64) -> f64 {
 
 /// Apply time decay to difficulty per share based elapsed time since last change,
 /// the interval we are decaying for and the incoming new difficulty value
-pub fn decay_time(dsps: f64, difficulty: u64, elapsed_time: f64, interval: u64) -> f64 {
+pub fn decay_time(dsps: f64, difficulty: f64, elapsed_time: f64, interval: u64) -> f64 {
     // Handle edge cases that could cause NaN or Infinity
     if elapsed_time <= 0.0 || !elapsed_time.is_finite() {
         return dsps; // Return the current value unchanged for invalid elapsed_time
